+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - WPF: Added `PageLifecycle`
 - WPF: Added `IWindowManager.ShowOpenFolderDialog(settings, ownerViewModel)`
 - WinUI: Added `WindowLifecycle`, `PopupLifecycle` and `PageLifecycle`
+- WinUI: Added `WindowManager`
 
 ### Changed
 - .NET 8.0 only